+++ TvDemo/nbproject/private/private.properties
@@ -11,6 +11,6 @@ user.properties.file=C:\\Users\\Luis Adrian\\AppData\\Roaming\\NetBeans\\8.0.2\\
 <<<<<<< HEAD
 user.properties.file=C:\\Users\\Windows\\AppData\\Roaming\\NetBeans\\12.1\\build.properties
 =======
-user.properties.file=C:\\Users\\DELL\\AppData\\Roaming\\NetBeans\\13\\build.properties
+user.properties.file=C:\\Users\\willi\\AppData\\Roaming\\NetBeans\\13\\build.properties
 >>>>>>> master
 >>>>>>> master
